+++ frontend/app/routes.ts
@@ -62,7 +62,9 @@ export default [
     ),
 
     route("settings", "./routes/layouts/settings-layout.tsx", [
-      index("./routes/settings/account-settings.tsx"),
+      index("./routes/settings/settings-index.tsx"),
+      route("account", "./routes/settings/account-settings.tsx"),
+      route("account/change-password", "./routes/settings/change-password.tsx"),
       route("ssh-keys", "./routes/settings/ssh-keys-list.tsx"),
       route("ssh-keys/new", "./routes/settings/create-ssh-key.tsx"),
       route("server-console", "./routes/settings/server-terminal.tsx"),
