+++ frontend/app/routes/settings/settings-index.tsx
@@ -1,0 +1,9 @@
+import { redirect } from "react-router";
+
+export function clientLoader() {
+  throw redirect("/settings/account");
+}
+
+export default function SettingsIndexPage() {
+  return null;
+}
